+++ changelog.md
@@ -1,5 +1,6 @@
 # 1.1.2.0
 
+* Fix `setBit`, `clearBit`, `complementBit` to preserve vector's length.
 * Fix Cabal 3.7+ incompatibility.
 
 # 1.1.1.0